# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Univibe API",
        "description": "API documentation for Univibe project",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/instructors": {
            "post": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Store a new instructor",
                "description": "Stores a new instructor and returns the created instructor",
                "operationId": "eafa861e61374164ae9f5395c66d6657",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "university_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Instructor created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/instructors/{id}": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Show a specific instructor",
                "description": "Returns a single instructor by ID",
                "operationId": "8535ef66d60a0f1728115c4869200929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the instructor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Update an existing instructor",
                "description": "Updates an existing instructor and returns the updated instructor",
                "operationId": "37dfc86f68847bc6236a645ff5164a55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the instructor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.7
                                    },
                                    "university_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Instructor updated successfully"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Delete an instructor",
                "description": "Deletes an existing instructor by ID",
                "operationId": "319535f0b03fcd732b98b460873db9c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the instructor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instructor deleted successfully"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}